finish him

+++ ft_printf.h
@@ -16,9 +16,10 @@
 # include <unistd.h>
 # include <stdlib.h>
 # include <stdarg.h>
+# include <stdio.h>
 
 int	ft_printf(const char *str, ...);
-size_t	ft_strlen(const char *str);
+int	ft_strlen(const char *str);
 int	ft_putnbr(int nbr, char *base, char c);
 int ft_putnbr_base(unsigned long int nbr, char *base);
 int ft_pointnbr(unsigned long int nbr, char *base);
